+++ matchms/utils.py
@@ -1,4 +1,5 @@
 import re
+from typing import Iterable
 from typing import Optional
 import numpy
 from .importing import load_adducts_dict
@@ -303,3 +304,10 @@ def clean_adduct(adduct: str) -> str:
 
     adduct_cleaned = adduct_core[:-len(adduct_charge)] + "]" + adduct_charge
     return adduct_conversion(adduct_cleaned)
+
+
+def get_first_common_element(first: Iterable[str], second: Iterable[str]) -> str:
+    """ Get first common element from two lists.
+    Returns 'None' if there are no common elements.
+    """
+    return next((item for item in first if item.lower() in second or item in second), None)